Fix: another bug related to data type

+++ src/models/arcf_with_embedding.py
@@ -76,8 +76,8 @@ def train(rating_generator, missing_generator, rating_discriminator,
             d_loss = Variable(torch.tensor(0, dtype=torch.float32, device=device), requires_grad=True)
             for i, batch in enumerate(train_dataloader):
                 review_embedding, real_rating_vector, index_item = batch
-                real_rating_vector = real_rating_vector.to(device)
-                review_embedding = review_embedding.to(device)
+                real_rating_vector = real_rating_vector.type(torch.float32).to(device)
+                review_embedding = review_embedding.type(torch.float32).to(device)
                 real_missing_vector = torch.tensor((real_rating_vector > 0) * 1).type(torch.float32).to(device)
                 index_item = index_item.type(torch.long).to(device)
                 noise_vector = torch.tensor(np.random.normal(0, 1, noise_size).reshape(1, noise_size),
